Improve/Fix controllers
